fluence-backend/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
pydantic-settings
google-adk
google-generativeai
google-cloud-aiplatform
google-cloud-firestore
google-cloud-storage
google-cloud-texttospeech
vertexai
httpx
python-multipart
google-cloud-texttospeech
